Param Converter Passes Original Guzzle Exception
When the GuzzleParamConverter finds a 404 error, the bundle throws a
NotFoundHttpException. This commit modifies that thrown exception to
also include with it the original Guzzle Exception.

This allows listeners of kernal.exception to check the
NotFoundHttpException for any BadResponseExceptions and work with the
additional information in the Guzzle request or response objects
attached to the BadResponseException.